refactor(Car): 자동차 이름 문자열 포장 클래스 CarName 생성
- 이름 관련 처리 이동
- carNameTooLongTest 테스트 케이스 이동
