+++ src/main/java/racingcar/model/Car.java
@@ -1,23 +1,17 @@
 package racingcar.model;
 
-import static racingcar.common.ErrorMessage.NAME_TO_LONG;
-
 public class Car {
-    private static final int NAME_MAX_LENGTH = 5;
-
-    private final String name;
+    private final CarName name;
 
     private Car(String name) {
-        this.name = name;
+        this.name = new CarName(name);
     }
 
     public static Car create(String name) {
-        if(name.length()>NAME_MAX_LENGTH)
-            throw new IllegalArgumentException(NAME_TO_LONG.text());
         return new Car(name);
     }
 
     public String getName() {
-        return this.name;
+        return this.name.getName();
     }
 }
